Makefile: Use switchers generated from master
For every other docs branch we can run set_versions.py once and have the correct
information in switchers.js. For transitions, we have multiple runs of the
documentation generation and we want to customise switchers.js in each case.
This is so we can show the current version, and the current release series
latest version which allows the switcher to work properly.  To do that, it is
easiest just to fix the transitions branch to handle that.

This allows an improved and custmomised switcher for each release showing the
current doc version, the latest in that series and the latest releases for
the project overall.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>